move retrieving nearby favors to mapPage
